methods.html: share redundant styles & pics with intro.html, and clean up headings.
Use the style sheet and the background image in intro_files/...
for both intro.html and methods.html.  Delete methods_files/...

Change title and headings in methods.html to use the same
pattern of HTML and styles as intro.html.

Change HTML <title> elements to "Editor's Introduction--Part X" (X = 1 or 2)

Make the PART 1 and PART 2 subheadings bold.

Reduce the letting in the subtitles from 5 to 3 pts.

Use the same body margins in both HTML files.